+++ src/components/Listbox/AsyncCreatableListbox.tsx
@@ -1,9 +1,8 @@
 import cx from "classnames";
-import Select, {
-  CommonProps as CommonSelectProps,
-  GroupBase,
-  Props as SelectProps,
-} from "react-select";
+import { GroupBase } from "react-select";
+import AsyncCreatableSelect, {
+  AsyncCreatableProps,
+} from "react-select/async-creatable";
 import { OmitStrict } from "../../utils/types";
 import { Icon } from "../Icon";
 import {
@@ -13,31 +12,32 @@ import {
   listBoxOnChange,
   ListboxOption,
 } from "./common";
-import "./Listbox.scss";
 import { ListboxBase, ListboxBaseProps } from "./ListboxBase";
 
-export type ListboxProps<Data, IsMulti extends boolean = boolean, Form = any> =
-  OmitStrict<
-    Partial<
-      // TODO: handle this any
-      CommonSelectProps<ListboxOption<Data>, IsMulti, any> &
-        SelectProps<
-          ListboxOption<Data>,
-          IsMulti,
-          GroupBase<ListboxOption<Data>>
-        >
-    >,
-    "onChange"
-  > &
-    ListboxBaseProps<Data, IsMulti, Form>;
+export type AsyncCreatableListboxProps<
+  Data,
+  IsMulti extends boolean = boolean,
+  Form = any,
+> = OmitStrict<
+  Partial<
+    AsyncCreatableProps<
+      ListboxOption<Data>,
+      IsMulti,
+      GroupBase<ListboxOption<Data>>
+    >
+  >,
+  "onChange"
+> &
+  ListboxBaseProps<Data, IsMulti, Form>;
 
-export function Listbox<Data, IsMulti extends boolean = false, Form = any>(
-  props: ListboxProps<Data, IsMulti, Form>,
-) {
+export function AsyncCreatableListbox<
+  Data,
+  IsMulti extends boolean = false,
+  Form = any,
+>(props: AsyncCreatableListboxProps<Data, IsMulti, Form>) {
   const {
     label,
     optionList,
-    placeholder,
     disabled,
     onChange,
     formValidation,
@@ -53,7 +53,7 @@ export function Listbox<Data, IsMulti extends boolean = false, Form = any>(
   return (
     <ListboxBase {...props}>
       {(formOnChange) => (
-        <Select<ListboxOption<Data>, IsMulti>
+        <AsyncCreatableSelect<ListboxOption<Data>, IsMulti>
           isOptionDisabled={() => (disabled ? true : false)}
           noOptionsMessage={listboxNoOptionMessage}
           {...selectProps}