+++ CHANGELOG.md
@@ -5,13 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.11] - 2022-04-22
+## [1.0.12] - 2022-04-27
+
+### Added
+
+- Slider: Add `renderThumbValue` method
+- Listbox: Add more variants to support async and creatable
+
+## [1.0.11] - 2022-04-26
 
 ### Added
 
 - Slider (#16)
 
-### FIxed
+### Fixed
 
 - Input's `hasContent` and active behavior (#12)